Mode management was moved to server worker.
