+++ src/routes/nodeinfo.js
@@ -8,23 +8,42 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 */
 
 import express from 'express';
+import bodyParser from '../middlewares/body_parser';
+import ensureJsonMime from '../middlewares/mime_middleware';
 import asyncMiddleware from '../middlewares/async_middleware';
 
-export const getNodeInfoHandler = (modelEngine, identityManager, gitCommit) => async (req, res) => {
+export const getNodeInfoHandler = (modelEngine, identityManager, gitCommit, operationalMode) => async (req, res) => {
   const workerLogs = await modelEngine.getWorkerLogs();
   require('pkginfo')(module, 'version');
   res.status(200).send({
     commit: gitCommit,
     version: module.exports.version,
+    mode: await operationalMode.get(),
     nodeAddress: identityManager.nodeAddress(),
     workerLogs
   });
 };
 
-const nodeInfoRouter = (modelEngine, identityManager, gitCommit) => {
+export const setModeHandler = (operationalMode) => async (req, res) => {
+  try {
+    await operationalMode.set(req.body);
+    res.status(200).send();
+  } catch (err) {
+    res.status(401)
+      .type('json')
+      .send();
+  }
+};
+
+const nodeInfoRouter = (modelEngine, identityManager, gitCommit, config, operationalMode) => {
   const router = new express.Router();
   router.get('/',
-    asyncMiddleware(getNodeInfoHandler(modelEngine, identityManager, gitCommit))
+    asyncMiddleware(getNodeInfoHandler(modelEngine, identityManager, gitCommit, operationalMode))
+  );
+  router.post('/',
+    ensureJsonMime,
+    bodyParser(config),
+    setModeHandler(operationalMode)
   );
   return router;
 };
